Moved PyTest fixtures in the "conftest.py" file.
